+++ src/2022-Oct-02/main-5.cpp
@@ -1,8 +1,8 @@
 #include <cppminimal>
 
-auto multiplu(std::span<int> a, int n, int k) -> int {
+auto multiplu(std::span<int> a, size_t n, int k) -> int {
     int cont = 0;
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         if ((a[i] % k == 0) && (a[i] % 10 == k)) cont++;
     }
